restructured interfaces
