+++ actions/add_plant.py
@@ -1,0 +1,2 @@
+def add_plant():
+    pass
