+++ actions/feed_animal.py
@@ -1,0 +1,2 @@
+def feed_animal():
+    pass
