+++ actions/report.py
@@ -1,5 +1,14 @@
-def build_facility_report(arboretum):
-    for river in arboretum.rivers:
-        print(f'River [{river.id}]')
+# def build_facility_report(arboretum):
+#     for river in arboretum.rivers:
+#         print(f'River [{river.id}]')
+    
+#     for mountains in arboretum.mountains:
+#         print(f'River [{river.id}]')
+    
+#     for coastline in arboretum.coastline:
+#         print(f'River [{river.id}]')
+    
+#     for forect in arboretum.rivers:
+#         print(f'River [{river.id}]')
 
-    input("\n\nPress any key to continue...")
+#     input("\n\nPress any key to continue...")
