+++ animals/river_dolphin.py
@@ -1,8 +1,10 @@
 from animals import Animal
-from interfaces.animal import IFreshwater
 from interfaces import Identifiable
+from interfaces import ISwimming
+from interfaces import ISaltwater
+from interfaces import IFreshwater
 
-class RiverDolphin(Animal, IFreshwater, Identifiable):
+class RiverDolphin(Animal, IFreshwater, ISaltwater, ISwimming, Identifiable):
 
     def __init__(self):
         Animal.__init__(self, "River dolphin")
