+++ index.py
@@ -1,8 +1,10 @@
 import os
 from arboretum import Arboretum
-from actions.annex import annex_habitat
-from actions.release_animal import release_animal
-from actions.report import build_facility_report
+from actions import annex_habitat
+from actions import release_animal
+from actions import build_facility_report
+from actions import add_plant
+from actions import feed_animal
 
 keahua = Arboretum("Keahua Arboretum", "123 Paukauila Lane")
