+++ interfaces/__init__.py
@@ -1,9 +1,14 @@
-from .terrestrial import ITerrestrial
-from .animal.freshwater import IFreshwater
-from .aquatic import IAquatic
-from .flying import IFlying
-from .walking import IWalking
-from .swimming import ISwimming
-from .identifiable import Identifiable
-from .habitat.contains_animals import IContainsAnimals
-from .habitat.contains_plants import IContainsPlants
+from .animal_interfaces.flying import IFlying
+from .animal_interfaces.freshwater import IFreshwater
+from .animal_interfaces.identifiable import Identifiable
+from .animal_interfaces.saltwater import ISaltwater
+from .animal_interfaces.swimming import ISwimming
+from .animal_interfaces.can_inhabit_forest import ICan_Inhabit_Forest
+from .animal_interfaces.can_inhabit_swamp import ICan_Inhabit_Swamp
+
+from .plant_interfaces.can_inhabit_forest import ICan_Inhabit_Forest
+from .plant_interfaces.can_inhabit_grassland import ICan_Inhabit_Grassland
+from .plant_interfaces.can_inhabit_mountain import ICan_Inhabit_Mountain
+from .plant_interfaces.can_inhabit_swamp import ICan_Inhabit_Swamp
+
+
