--- interfaces/animal/__init__.py
@@ -1,1 +1,0 @@
-from .freshwater import IFreshwater
