+++ interfaces/animal_interfaces/can_inhabit_forest.py
@@ -1,0 +1,4 @@
+class ICan_Inhabit_Forest:
+
+    def __init__(self):
+       pass
