+++ interfaces/animal_interfaces/flying.py
@@ -1,0 +1,7 @@
+class IFlying:
+
+    def __init__(self, feathers, wing_count=2):
+        self.feathers = feathers
+        self.wing_count = wing_count
+
+
