+++ interfaces/animal_interfaces/freshwater.py
@@ -1,0 +1,7 @@
+
+class IFreshwater:
+
+   def __init__(self):
+        self.cell_type = "hypotonic"
+
+
