+++ interfaces/animal_interfaces/saltwater.py
@@ -1,0 +1,4 @@
+class ISaltwater():
+
+    def __init__(self):
+        self.cell_type = "hypertonic"
