+++ interfaces/animal_interfaces/swimming.py
@@ -1,0 +1,4 @@
+class ISwimming:
+
+    def __init__(self):
+        self.fins = True
