--- interfaces/aquatic.py
@@ -1,4 +1,0 @@
-class IAquatic:
-
-    def __init__(self):
-        self.aquatic = True
