--- interfaces/habitat/contains_animals.py
@@ -1,4 +1,0 @@
-class IContainsAnimals():
-
-    def __init__(self):
-        self.animals = []
