--- interfaces/habitat/contains_plants.py
@@ -1,4 +1,0 @@
-class IContainsPlants():
-
-    def __init__(self):
-        self.plants = []
