+++ interfaces/plant_interfaces/can_inhabit_grassland.py
@@ -1,0 +1,4 @@
+class ICan_Inhabit_Grassland:
+
+    def __init__(self):
+        pass
