+++ interfaces/plant_interfaces/can_inhabit_mountain.py
@@ -1,0 +1,4 @@
+class ICan_Inhabit_Mountain:
+
+    def __init__(self):
+        pass
