+++ interfaces/plant_interfaces/can_inhabit_swamp.py
@@ -1,0 +1,4 @@
+class ICan_Inhabit_Swamp:
+
+    def __init__(self):
+        pass
