--- interfaces/saltwater.py
@@ -1,5 +1,0 @@
-class ISaltwater():
-
-    def __init__(self):
-        super().__init__()
-        self.cell_type = "hypotonic"
