--- interfaces/swimming.py
@@ -1,8 +1,0 @@
-class ISwimming:
-
-    def __init__(self):
-        self.swim_speed = 0
-        self.maximum_depth = 0
-
-    def swim(self):
-        print("Glug glug, I'm swimming!")
