--- interfaces/terrestrial.py
@@ -1,4 +1,0 @@
-class ITerrestrial:
-
-    def __init__(self):
-        self.terrestrial = True
