--- interfaces/walking.py
@@ -1,5 +1,0 @@
-class IWalking:
-
-    def __init__(self, leg_count=2):
-        self.leg_count = leg_count
-        self.move_speed = 0
